responsable, microempresa,familia

+++ data/responsableData.php
@@ -1,8 +1,8 @@
 <?php
-    
+
     include 'data.php';
     include '../domain/responsable.php';
-    
+
     class responsableData{
 
         public function responsableData(){}
@@ -12,11 +12,11 @@
 
             $con = new Data();
             $conexion = $con->conect();
-            
+
 
         $cedula = $responsable->getCedulaResponsable();
         $nombre = $responsable->getNombreResponsable();
-        $primerApellido = $responsable->getPrimirapellidoResponsable();   
+        $primerApellido = $responsable->getPrimirapellidoResponsable();
         $segundoApellido = $responsable->getSegundoapellidoResponsable();
         $telefono = $responsable->getTelefonoResponsable();
         $email = $responsable->getEmailResponsable();
@@ -32,7 +32,7 @@
             }
 
             $consultaInsertar="INSERT INTO tbresponsable VALUES (
-            
+
             ".$idSiguiente.",
             '".$cedula."',
             '".$nombre."',
@@ -68,12 +68,12 @@
             while ($row = mysqli_fetch_array($result)) {
                 $temporaralResponsable = new Responsable(
 
-                    $row['idresponsable'], 
-                    $row['cedularesponsable'], 
-                    $row['nombreresponsable'], 
+                    $row['idresponsable'],
+                    $row['cedularesponsable'],
+                    $row['nombreresponsable'],
                     $row['primerapellidoresponsable'],
                     $row['segundoapellidoresponsable'],
-                    $row['telefonoresponsable'], 
+                    $row['telefonoresponsable'],
                     $row['emailresponsable']
                 );
 
@@ -93,23 +93,46 @@
             $con = new Data();
             $conexion = $con->conect();
 
-            $consultaActualizar = "UPDATE tbresponsable SET 
-
-        cedularesponsable = '".$responsable->getCedulaResponsable()."',
-        nombreresponsable = '".$responsable->getNombreResponsable()."',
-        primerapellidoresponsable = '".$responsable->getPrimirapellidoResponsable()."',
-        segundoapellidoresponsable = '".$responsable->getSegundoapellidoResponsable()."',
-        telefonoresponsable = '".$responsable->getTelefonoResponsable()."',
-        emailresponsable = '".$responsable->getEmailResponsable()."' 
-        WHERE idresponsable =" . $responsable->getIdResponsable() . ";";
 
 
+        $consultaFamilia ="SELECT idresponsable FROM tbfamilia where idresponsable=". $responsable->getIdResponsable() ."";
+       $idFamilia=mysqli_query($conexion,$consultaFamilia);
+        if ($row = mysqli_fetch_row($idFamilia)) {
+                $idTemporal = trim($row[0]);
+            }
 
+            if($idTemporal==$responsable->getIdResponsable()){
+              $retorno=1;
+            }else{
+
+              $consultaEmpresa ="SELECT idresponsable  FROM tbmicroempresa where idresponsable =".$responsable->getIdResponsable()."";
+          $idEmpresa=mysqli_query($conexion,$consultaEmpresa);
+
+          if ($row = mysqli_fetch_row($idEmpresa)) {
+                  $idTemporal = trim($row[0]);
+              }
+
+              if($idTemporal==$responsable->getIdResponsable()){
+                $retorno=2;
+              }else{
+
+                $consultaActualizar = "UPDATE tbresponsable SET
+
+            cedularesponsable = '".$responsable->getCedulaResponsable()."',
+            nombreresponsable = '".$responsable->getNombreResponsable()."',
+            primerapellidoresponsable = '".$responsable->getPrimirapellidoResponsable()."',
+            segundoapellidoresponsable = '".$responsable->getSegundoapellidoResponsable()."',
+            telefonoresponsable = '".$responsable->getTelefonoResponsable()."',
+            emailresponsable = '".$responsable->getEmailResponsable()."'
+            WHERE idresponsable =" . $responsable->getIdResponsable() . ";";
             $result = mysqli_query($conexion,$consultaActualizar);
             mysqli_close($conexion);
 
-            return $result;
+
         }
+      }
+        return $retorno;
+    }
 
 
 
@@ -118,18 +141,39 @@
 
 
 
-        public function eliminarResponsable($idresponsable)
-        {
+        public function eliminarResponsable($idresponsable){
+
             $con = new Data();
             $conexion = $con->conect();
 
-            $consultaEliminar="DELETE FROM tbresponsable WHERE idresponsable =".$idresponsable.";";
+            $consultaFamilia ="SELECT idresponsable FROM tbfamilia where idresponsable=".$idresponsable ."";
+           $idFamilia=mysqli_query($conexion,$consultaFamilia);
+            if ($row = mysqli_fetch_row($idFamilia)) {
+                    $idTemporal = trim($row[0]);
+                }
 
-            $result=mysqli_query($conexion,$consultaEliminar);
-            mysqli_close($conexion);
+                if($idTemporal==$idresponsable){
+                  $retorno=1;
+                }else{
 
-            return $result;
-        }
+                  $consultaEmpresa ="SELECT idresponsable  FROM tbmicroempresa where idresponsable =".$idresponsable."";
+                  $idEmpresa=mysqli_query($conexion,$consultaEmpresa);
 
+                  if ($row = mysqli_fetch_row($idEmpresa)) {
+                      $idTemporal = trim($row[0]);
+                  }
+
+                  if($idTemporal==$idresponsable){
+                    $retorno=2;
+                  }else{
+
+                      $consultaEliminar="DELETE FROM tbresponsable WHERE idresponsable =".$idresponsable.";";
+
+                      $result=mysqli_query($conexion,$consultaEliminar);
+                      mysqli_close($conexion);
+                    }
     }
+     return $retorno;
+  }
+}
   ?>
